(TKP-7) Moved post request to wasteReport
